backend: limitacion y actualizacion archivos
Se limita el borrado de archivos, de modo que si hay 3 o menos archivos en el curso, no se puede borrar un archivo, de modo que debe actualizarlo con la nueva funcionalidad implementada. Adicionalmente, se añade un nuevo término al apartado de terminos.

+++ app/views.py
@@ -1,8 +1,9 @@
 from django.core.files.storage import default_storage, DefaultStorage, FileSystemStorage
 from django.core.files.base import ContentFile
 from django.shortcuts import render, redirect
+from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
-from django.http import JsonResponse
+from django.http import JsonResponse, HttpResponseRedirect
 from django.urls import reverse
 from app import paypal
 from app.forms import MonederoForm, RetiradaDineroForm, UsuarioForm, CursoForm, ReporteForm, UploadFileForm, CursoEditForm, ActualizarUsuarioForm, ComentarioForm, ResponderComentarioForm, ResponderComentarioForm2
@@ -677,6 +678,7 @@ def curso(request, id, suscrito=False):
         if usuario.titulacion != curso.asignatura.titulacion and usuario not in curso.suscriptores.all() and usuario != curso.propietario:
             return redirect('/cursosdisponibles')
         form = UploadFileForm(request.POST, request.FILES)
+        formUpdate = UploadFileForm(request.POST, request.FILES)
         excede_tamano = False
         excede_mensaje = ""
         valoracionCurso = get_valoracion(curso)
@@ -701,6 +703,8 @@ def curso(request, id, suscrito=False):
                             excede_mensaje = e.message_dict['ruta'][0]
                 else:
                     form = UploadFileForm()
+                    formUpdate = UploadFileForm()
+                
         elif usuario in curso.suscriptores.all():
             es_suscriptor = True
             try:
@@ -709,7 +713,7 @@ def curso(request, id, suscrito=False):
             except:
                 pass
 
-        return render(request, "curso.html", {"id": id, "es_owner": es_owner, "es_suscriptor": es_suscriptor, "curso": curso, "contenido_curso": contenido_curso, "form": form, "excede_tamano": excede_tamano, "excede_mensaje": excede_mensaje, "valoracionCurso": valoracionCurso, "valoracionUsuario": valoracionUsuario, "suscrito": suscrito,"nombre_archivo_unico": nombre_archivo_unico})
+        return render(request, "curso.html", {"id": id, "es_owner": es_owner, "es_suscriptor": es_suscriptor, "curso": curso, "contenido_curso": contenido_curso, "form": form, "formUpdate": formUpdate, "excede_tamano": excede_tamano, "excede_mensaje": excede_mensaje, "valoracionCurso": valoracionCurso, "valoracionUsuario": valoracionUsuario, "suscrito": suscrito,"nombre_archivo_unico": nombre_archivo_unico})
 
     else:
         return render(request, 'inicio.html')
@@ -745,6 +749,10 @@ def borrar_archivo(request, id_curso, id_archivo):
         usuario_autenticado = request.user
         usuario = Usuario.objects.get(django_user=usuario_autenticado)
         if (curso.propietario == usuario):
+            numero_archivos_curso = curso.archivos.count()
+            if(numero_archivos_curso < 4):
+                messages.info(request, 'Mínimo de archivos posible - Debes tener al menos 3 archivos subidos a un curso para poder borrar uno, prueba a actualizarlo!')
+                return HttpResponseRedirect('/curso/' + str(id_curso))
             archivo = Archivo.objects.get(id=id_archivo)
             archivo.delete()
             nombre = archivo.nombre.replace(" ", "_")
@@ -1031,6 +1039,45 @@ def servir_archivo(request, id_curso, archivo):
             return error_403(request, None)
     else:
         return error_403(request, None)
+    
+def editar_archivo(request, id_curso, id_archivo):
+    curso = Curso.objects.get(id=id_curso)
+    if request.user.is_authenticated:
+        # Comprobar si el usuario es profesor
+        usuario_autenticado = request.user
+        usuario = Usuario.objects.get(django_user=usuario_autenticado)
+        formUpdate = UploadFileForm(request.POST, request.FILES)
+        if (curso.propietario == usuario):
+            if request.method == 'POST':
+                if formUpdate.is_valid():
+                    file = request.FILES['file']
+                    archivo_instancia = Archivo(
+                        nombre=file.name, ruta=file, curso=curso)
+                    if file.name in curso.archivos.values_list('nombre', flat=True):
+                        messages.info(request, 'Archivo repetido - El archivo ya existe en el curso')
+                        return HttpResponseRedirect('/curso/' + str(id_curso))
+                    else:
+                        #añadir archivo
+                        try:
+                            archivo_instancia.full_clean()
+                            #borrar archivo
+                            archivo = Archivo.objects.get(id=id_archivo)
+                            archivo.delete()
+                            nombre = archivo.nombre.replace(" ", "_")
+                            try:   
+                                os.remove("files/" +
+                                    str(curso.id) +"/" +  nombre)
+                                print("Archivo borrado")
+                            except:
+                                print("No se pudo borrar el archivo")
+                                messages.info(request, 'Error al subir el archivo- Prueba de nuevo en unos minutos')
+                                return HttpResponseRedirect('/curso/' + str(id_curso))
+                            archivo_instancia.save()
+                        except ValidationError as e:
+                            excede_mensaje = e.message_dict['ruta'][0]
+                            messages.info(request, excede_mensaje)
+                            return HttpResponseRedirect('/curso/' + str(id_curso))
+    return redirect('/curso/'+str(id_curso))
 
 
 def sobre_nosotros(request):
